delete user

+++ BlockstackUser.php
@@ -70,10 +70,15 @@ class BlockstackUser {
 		}
 	}
 
+	/**
+	 * Remove this Blockstack user from the table
+	 */
 	public function remove() {
-		$dbw = wfGetDB( DB_MASTER );
-		$dbw->delete( BlockstackSso::TABLENAME, ['bs_did' => $this->did] );
-		$this->exists = false;
+		if( $this->exists ) {
+			$dbw = wfGetDB( DB_MASTER );
+			$dbw->delete( BlockstackSso::TABLENAME, ['bs_did' => $this->did] );
+			$this->exists = false;
+		}
 	}
 
 	public function getDid() {
